feat: makes slug update once the base field update (#13)
The previous implementation was not updating the slug content when the
base slug field were updated. This change improve this behavior, making
the slug generation regardless the model operation: create or update.
This may not be an issue, since the slug right now is not planned to be
a long-term key for the model.