feat: add LandscapeGroup GraphQL mutations (#29)
* feat: adds LandscapeGroup query to the GraphQL API

* feat: add query for Membership and GroupAssociation

This change adds on GraphQL API the queries for Membership and
GroupAssociation models.

* refactor: move relationship GraphQL schema to proper modules

This change is necessary because the relationship models themselves will
have mutations and it would increase too much the modules where they
were.

* feat: adds LandscapeGroup query to the GraphQL API

* refactor: move relationship GraphQL schema to proper modules

This change is necessary because the relationship models themselves will
have mutations and it would increase too much the modules where they
were.

* refactor: make base delete work with compound model names

This change updates the base delete mutation to make it properly work
with models that have compound names, like LandscapeGroup. In these
cases, the result attribute name is used in snake case, like
landscape_group.

* refactor: removes duplication of client_query fixture

* feat: add LandscapeGroup GraphQL mutations

* refactor: compile regex to convert camel to snake case

This is necessary to optimize regex execution.