fix: flaky Membership mutation test (#37)
Before this change, it was being possible to have more then one
membership for the same user in a given group. This was causing the
delete mutation test randomly fail, not because of process failure, but
because the way the result were being checked. This change makes sure
only one user will be present on each group for GraphQL tests.

This is the quick win fix. Since there's a possibility of implementing
soft delete (so, changing the way to check object removal), let's move
on with this implementation at the moment.