feat: call model validation on GraphQL mutations (#34)
This change calls for Django model validation before database writing
during the GraphQL mutations (add and update). This implementation was
done having in mind the GraphQL specs about response errors.

According to the GraphQL Spec[1]:
> A response to a GraphQL request must be a map.
> If the request raised any errors, the response map must contain an
> entry with key errors.

Taking this into consideration and the only way to return errors on
response supported by graphene-django, this implementation of model
validation have to raise an exception during the mutation execution.
This exception is going to be catch and handled by the graphene-django
view and the proper errors element will be added to the response
payload.

A custom GraphQL validation exception were created both to have a
semantic error, but also to implement the logic of extracting the error
messages from the Django ValidationError and putting into one single
message. Currently, graphene-django supports only one error message per
response.

[1] - http://spec.graphql.org/October2021/#sec-Response-Format