feat: add command to import Landscape data (#33)
* feat: add command to import Landscape data

This change adds a custom admin command to the core app to import
Landscape (and Group) data from AirTable. The import is made accessing
AirTable API. So, in order to execute this command, a API key is
necessary. The API key can be generated on AirTable account settings.
The API key might be informed as a command argument or it also can be
set as an environment variable called AIRTABLE_API_KEY.

The management command can be run as:

```sh
python terraso_backend/manage.py loadsampledata
python terraso_backend/manage.py loadsampledata --airtable_api_key=XXX
```

* feat: add sample data extracted from AirTable

This change adds a sample data file to be used with Django's `loaddata`
command. It also adds instructions on README about how to run it.