feat: implements authentication based on JWT (#59)
This change adds the support to user authentication based on JWT tokens.
To make an authenticated request, the client must inform the token on
`AUTHORIZATION` header following the format `Bearer <token>`.

**Important**: after this change, it will still be allowed to perform
API calls without being authenticated. The restriction will be activated
on further changes. This way we can smoothly deploy this change, add
support for authentication to the web client, then synchronously
activate authentication on API and client.

With this feature, after the user signs up with Apple or Google, the web
client will receive a redirect response from the API with a cookie. This
cookie value will be a JSON with both access and refresh tokens.

The access token is the one required to perform authenticated requests.
Both tokens can be handled as _opaque_ tokens. What means that it only
has meaningful value for the back-end.

To fetch the authenticated user information, the client can send an
authenticated GET request to `/auth/user`.

Both access and refresh tokens has configurable expiring delta times
that can be set through environment variables.

Once the access token is expired, a new pair of tokens (access and refresh)
can be requested on `POST /auth/tokens` informing a previous valid refresh
token on request body (JSON).