refactor: creates AuthenticationRequiredMixin to auth check
This change creates a mixin to centralize the check if user is logged in
or not. It also properly build the response object with the expected
status code. It's worth to see that for requests with authentication
failure, the response will always be a JSON response with a proper HTTP
status code (401).

The added mixin is intended to be used instead of multiplying across the
code base the verification `if not user.is_authenticated`. So, on every
view that is making this check, the `AuthenticationRequiredMixin` should
be used instead.

The Django's `LoginRequiredMixin` wasn't used because:

1 - It returns 403 responses for non authenticated requests;
2 - There are only redirect or HTML responses in case of authentication
  failure and we need a JSON response since it's more convenient for our
  use case.

As any other mixin, the `AuthenticationRequiredMixin` should be at the
leftmost position in the inheritance list.