feat: Add created_by field to Landscape and Group
This change adds a new attribute named `created_by` on Landscape and
Group models. This field is going to be used to store the information of
which user was the creator of that entity. This might me useful for
future auditory and also for future automation, like automatically
creating a Membership for the Group creator.