feat: makes addMembership mutation idempotent
This change updates the `addMembership` mutation on GraphQL to be
idempotent. It means that, if the same pair of `user` and `group` be
submitted to membership creation more than once, only one membership
will be created.

This change is necessary in a context where the web client is having to
create the Group manager directly, because the back-end hadn't this
automatic creation implemented.

Now that the back-end implements the automatic manager creation, we have
to create the conditions to release it without breaking the current
implementation of web client.

So, after this change, we can release the automatic group manager
creation on the API side without the need to sync the web client update.