fix: user being updated on every sign in
This change updates the Account Service to make sure the user name is
only set when user is created. If the user signing in was already
created on database, then their names will be left as is. This is
important because users might have had updated their names on Terraso
platform after the sign up process.