feat: add created_by attribute to Landscape and Group  (#85)
* feat: Add created_by field to Landscape and Group

This change adds a new attribute named `created_by` on Landscape and
Group models. This field is going to be used to store the information of
which user was the creator of that entity. This might me useful for
future auditory and also for future automation, like automatically
creating a Membership for the Group creator.

* feat: add Group and Landscape created_by on GraphQL

This change updates schema and mutations (internally) on GraphQL to
support the attribute `created_by` on Group and Landscape models.