fix: way of checking a Group is being created
This change fix the way of checking if a Group is being created (on
Group.save()) to automatically create the manager Membership. It's
necessary because the model PK defines a default value (uuid). So, even
on creation call the pk field won't ever be false. This way, we have to
check if a model if that PK exists on database to determine if it's a
create or a update call.