feat: adds Group Manager creation when Group is created (#87)
* feat: group creator is automatically set as group manager

This change updates the `Group.save()` method to automatically create a
`Membership` as role manager for the group creator.

* feat: makes addMembership mutation idempotent

This change updates the `addMembership` mutation on GraphQL to be
idempotent. It means that, if the same pair of `user` and `group` be
submitted to membership creation more than once, only one membership
will be created.

This change is necessary in a context where the web client is having to
create the Group manager directly, because the back-end hadn't this
automatic creation implemented.

Now that the back-end implements the automatic manager creation, we have
to create the conditions to release it without breaking the current
implementation of web client.

So, after this change, we can release the automatic group manager
creation on the API side without the need to sync the web client update.

* fix: way of checking a Group is being created

This change fix the way of checking if a Group is being created (on
Group.save()) to automatically create the manager Membership. It's
necessary because the model PK defines a default value (uuid). So, even
on creation call the pk field won't ever be false. This way, we have to
check if a model if that PK exists on database to determine if it's a
create or a update call.