feat: allows only Group Manager update Group data (#93)
This change adds permissions check to allow only Users with Group
Manager role to update Group data.

The permissions check works based on `django-rules`[1] lib.

One of the core concepts of `django-rules` are the `predicates`.
Predicates are simple _callables_ that returns a boolean value as a
permission result. The way it was set on the project, all our predicates
should be defined only on `apps.core.permission_rules`.

The base models were updated to integrate with `django-rules` and to
make possible to associate the predicates directly on model definition
(Meta class). This way is somehow explicit which permissions are related
to each model and to which model operation (create, update, delete).

On any place where we need to check for permission before an operation,
we can do it calling an instruction similar to this base one:
`User.has_perm(Model.get_perm('...'), obj=model)`. This is going to call
the proper predicate associated to the model permission.

It's important to note that predicates (and permissions) can be created
regardless the models. Which means that the same foundation might be
used to other permissions checks that not necessarily involves CRUD
model operations. More details can be found on linked lib README.

This is part of:
- #70 

[1] - https://github.com/dfunckt/django-rules