fix: unique constraints with soft deleted objects
Because unique constraints are set at the database level, the
traditional ways of defining uniqueness doesn't work with the soft
delete solution in place. Besides, this can lead to confusion as the
soft deleted objects are not visible by the user. This is being solved
here by setting a partial unique constraint that will only check
uniqueness on non-deleted objects, as recommended by `django-safedelete`
lib[1].

[1] - https://github.com/makinacorpus/django-safedelete/blob/master/docs/models.rst#fields-uniqueness