fix: GraphQL Group filters using deleted associated objects
There was a specific case where using Django built-in lookups for
relationships doesn't work properly with the soft delete solution we
implemented with `django-safedelete`. According to the library
maintainers, it's a known limitation[1].

To work around this limitation, a customized filter set were created for
Group schema on GraphQL. This custom filter set implements the
adjustments expected to ignore soft deleted objects when filtering by
relationships (joins).

[1] - https://github.com/makinacorpus/django-safedelete/issues/158