chore: bump Django version to 4.0 (#114)
This change updates the Django version to 4.0.

This was only possible making a monkey patching of a specific function
removed in Django 4.0 but that is still used by the current
graphene-django version (2.15). This is a temporary measure only to do
not keep depending on the slow process on graphene-django updates. As
soon as a new version is released we can remove the workaround.