feat: add state parameter to authorize URL API
This change adds the `state` parameter to the `{provider}/authorize`
API. This endpoint returns the proper provider URL to redirect user to
authentication. Now, this API accepts a `state` parameter that will be
used on redirect back user to the client. Example:

If the authorize URL for Google be request like:

```
/auth/apple/authorize?state=groups/new
```

After a successful authentication, the user will be redirect to
`{web_client_endpoint}/groups/new`.