feat: add state parameter to authorize URL API (#116)
* chore: rename self to cls on classmethod

* feat: add state parameter to authorize URL API

This change adds the `state` parameter to the `{provider}/authorize`
API. This endpoint returns the proper provider URL to redirect the user to
authentication. Now, this API accepts a `state` parameter that will be
used to redirect back user to the client. Example:

If the authorize URL for Google, be request like:

```
/auth/apple/authorize?state=groups/new
```

After a successful authentication, the user will be redirected to
`{web_client_endpoint}/groups/new`.