feat: add error codes to validation error responses (#115)
This change updates how the validation messages are built. It now makes
the message be a serialized JSON in order to add more contextual data to
the message. This is an attempt to make possible the i18n of validation
messages by the clients.

This is an example of validation error message using the new schema:

```
'message': '[{
    "code": "updateNotAllowed",
    "context": {"model": null, "field": "membership"}
}]'
```

If the validation happens in a specific field, the context `field`
attribute will have the related field name. Otherwise, it will be null. 

```
'message': '[{
    "code": "notFound",
    "context": {"model": "Membership", "field": "group", "extra": ""}
}]'
```

In some specific cases, an error message might have a complement to the
error code. Like in a situation where a `Membership` cannot be updated
because the user is the last manager in that group:

```
'message': '[{
    "code": "updateNotAllowed",
    "context": {"model": "Membership", "field": null, "extra": "manager_count"}
}]'
```

In cases where more than one field have validation errors, then the
message will have more than one object:

```
'message': '[
    {"code": "blank", "context": {"model": "Landscape", "field": "name"}},
    {"code": "invalid", "context": {"model": "Landscape", "field": "website"}
}]'
```

To make this structure work, more specific exceptions were added to the
GraphQL app with the proper handler to format the error message. Also,
all mutations were already updated to work with the new schema.