fix: Membership update mutation
During mutations refactoring a call to `super()` was left on membership
update mutation process, but the mutations inheritance structure was
changed. This call was being made based on the value of
`CHECK_PERMISSIONS` feature flag and the test was not covering the bug
scenario. So, the test was also updated.