feat: add back-end session
In order to integrate Django admin authentication with our current
authentication system, we have to use Django sessions. Also, in near
future we will have to use Django sessions to enable users authenticate
to other services using OAuth.

So, this change increments our current authentication process to
properly create Django sessions after user signs in. The only difference
in the sign in process is the call to Django's `login()`.

Also, to provide a mechanism to end sessions, this change adds a new
endpoint to the API `/logout`. This view simply call Django's `logout()`
function and returns `OK` to the client.