fix: duplicated slug message returned by the API (#148)
This change fixes the response message for duplicated slug scenarios
when using the generic write mutation. As a good collateral effect, this
change already handle any integrity error that might happen.

The slug unique constraint only runs over non-deleted registers. This
constraint is created directly on DB. So, when performing a save
operation, this unique constraint will only be assessed on DB level, not
on application level. The DB error message is the message being returned
by the API at the moment.

This change handles the `IntegrityError` when calling `save()` and build
an appropriate `ValidationError` to be returned by the API.