feat: add upload view to create data entries (#166)
* chore: add missing config to env sample

* chore: add data entry s3 config

* chore: add python-magic and update dependencies

The python-magic lib will be used to validate data entry files type with
their extensions.

* feat: update base upload service

This change updates the base upload service. It makes upload method
accept a file name, and in case it's not informed a unique name is
created. It also adds the feature to upload file with the same name
adding a counter as the name suffix.

* feat: add upload view to create data entries

* fix: add test mock

* chore: remove unused save override

* chore: remove unused imports

* fix: only checks file type and extension if indentified

This change runs the check between file type and file extension only if
both (file type and file type from extension) could be identified.
Otherwise, it won't validate.

* chore: adds log calls to shared data forms processing

* fix: uniquify method

This change fix the uniquify method that was not building the new file
name with counter correctly.

* chore: improve error result for DataEntry upload API

* feat: changes Data Entry upload to work with Group slug

* chore: improve data file type and extension validation

* feat: add mailcap to the Docker image

This change adds mailcap to the Docker image. The mailcap adds more
mimetypes support to the system. This is needed to the validation
process of file type and extension on `DataEntry` upload.

* fix: data entry view test

* feat: fill out data entry size on upload process

* feat: use signed URL for DataEntry URL attribute

* chore: simplify data entry model tests extracting fixture

* chore: new line at the end of file

* chore: persist file extension instead of mimetype