fix: duplicated Data Entries on API response (#176)
In order to answer only `DataEntry` from groups that the current user is
member of, a specific filter is being done on GraphQL `DataEntryNode`.
Until this change the query was relying on built-in Django look-ups for
the joins. However, it was including soft-deleted registers (Membership
records) because of not so good integration between Django's look-ups
and the `django-safedelete` lib. This was causing the `DataEntry`
duplication in some groups.

To fix that issue the query was updated to be more explicit and done in
two steps:

1. Get all ID from Groups that the current user is member of. Since this
   is a simple query, the soft-deleted is well respected.
2. With the list of group IDs, filter the `DataEntry` query. This will
   also respect the soft-delete underlying conditions