feat: set cookie on signup so it can be tracked by the frontend
The frontend has to track signups when they occur. Unfortunately this
is not really possible right now because the OAuth callback is handled
on the backend. The idea here is that we can set the cookie when a new
user is created, the cookie can be read on the frontend, a request
will be sent to plausible to track which service was used to create
the user, and finally the cookie will be discarded.