feat: query plausible on backend instead of setting cookie
In the first attempt, a cookie was set when a new user signed up. This
cookie was then used by the frontend to tell Plausible that a new user
had been registered with a given service (right now Apple or Google).

This has been changed to completely eliminate the frontend
role. Plausible is now used on the backend. In order to do this, there
is a new class in the `apps/auth/services.py` file that can be used to
send the tracking event to Plausible via their HTTP API. This service
could be re-used in the future if we ever felt the need. For now I
just put it in the `auth` app because this is the only place where it
is used.