feat: Create script to run backend and S3 sync
This script does the following:

1) syncs the S3 buckets that need to be copied
2) backs up DB A to the backup S3 bucket
3) loads DB backup into DB B

It can be quite confusing, because the script is *not* designed to be
run in either environment A or B. You can think about it it as being
run in some environment C, which could just be your local machine. It
executes commands in environment A and B through the Render HTTP
API. It also uses the AWS cli to sync the S3 buckets.

The AWS cli needs to be installed in C. I provided a Dockerfile
as well here, but it will need to have the AWS cli credentials mounted
when it is run. It also needs two environment variables to be
provided.
    - $RENDER_TOKEN is the Render HTTP API Token
    - $URL_REWRITES_CONFIG_FILE is a file that should be stored in B
    and indicates the location of a config file that defines the URLs
    that need to be replaced in the database to point to the B
    environment S3 buckets. (This is documented elsewhere). The idea
    is this file can be provided as a "secret file" in the Render interface.