feat: Add function to calculate polygon area (#210)
This function takes the GeoJSON representation of a polygon and
returns the area. It is just a wrapper around a function from the
`pyproj` library, converting the GeoJSON to a `shapely` shape. It also
uses a default projection of WSG84, a standard projection that uses
lat/lon degrees. I believe this is the default for Leaflet, but I'll
have to take a close look.

The test tries to compute a 1m x 1m rectangle in a UTM CRS and then
transform the coordinates to WGS84. The rectangle's area is about
0.985 or so, which seems good to me. I'm not sure how to estimate the
possible error here, it might be interesting to try another test.