feat: Save landscape polygon area on landscape upate (#210) (#221)
* feat: Add function to calculate polygon area (#210)

This function takes the GeoJSON representation of a polygon and
returns the area. It is just a wrapper around a function from the
`pyproj` library, converting the GeoJSON to a `shapely` shape. It also
uses a default projection of WSG84, a standard projection that uses
lat/lon degrees. I believe this is the default for Leaflet, but I'll
have to take a close look.

The test tries to compute a 1m x 1m rectangle in a UTM CRS and then
transform the coordinates to WGS84. The rectangle's area is about
0.985 or so, which seems good to me. I'm not sure how to estimate the
possible error here, it might be interesting to try another test.

* feat: Calculate boundary area on saving Landscape (#210)

* feat: Add migrations for landscape boundary area (#211)

* fix: Change area_scalar to reflect units

* fix: Use GeoJSON default CRS for all boundaries

This is outlined in https://www.rfc-editor.org/rfc/rfc7946#section-4
. We should only support what the standard supports, do no need to do extra.

* feat: Add boundary area field to GraphQL schema (#210)

* fix: Add missing m2 to area storage (#210)

* feat: Add area in hectares to Landscape GraphQL Schema

* fix: Use default CRS instead of giving option

Because the current GeoJSON spec only uses a default CRS and removes
the ability to specify a custom CRS, we shouldn't support custom CRS either.

* feat: Add django-dirtyfields library

* feat: Only calculate landscape area if boundary has changed

* fix: Fix up migration conflict

* fix: Address PR comments

* fix: Calculate area for multiple GeoJSON features

I mistakenly thought that the landscape boundary was only a single
polygon, but this is not the case. The new function calculates area
for any collection of Polygon or MultiPolygon. Caveats: I didn't bother cover
GeometryCollection yet, and it doesn't take into account intersections.