feat: Landscape profile (#211)
* feat: Added landscape profile fields and taxonomy terms

* feat: Added taxonomy term type options

* feat: Added Taxonomy terms to graphql schema

* feat: Translatable taxonomy terms

* fix: Support multiple area types

* feat: Save taxonomy terms

* fix: Added development strategy to graphql schema, reuse taxonomy save logic

* fix: Change population to integer

* feat: Landscape Partnership relation, save group associations

* feat: Update landscape group associations

* fix: Fixed development stratey relationship

* fix: Added Landscape group to admin, fixed taxonomy term name

* feat: Added is partnership filter

* feat: Added email to Graph QL schema

* fix: Check if relationship fields exist

* chore: Added tests for landscape update relationships

* chore: Added Taxonomy term tests

* fix: Removed duplicated groups

* fix: Fixed quotes

* feat: Populate taxonomy terms (#216)

* feat: Upload Landscape Profile Image (#219)

* feat: Save landscape polygon area on landscape upate (#210) (#221)

* feat: Add function to calculate polygon area (#210)

This function takes the GeoJSON representation of a polygon and
returns the area. It is just a wrapper around a function from the
`pyproj` library, converting the GeoJSON to a `shapely` shape. It also
uses a default projection of WSG84, a standard projection that uses
lat/lon degrees. I believe this is the default for Leaflet, but I'll
have to take a close look.

The test tries to compute a 1m x 1m rectangle in a UTM CRS and then
transform the coordinates to WGS84. The rectangle's area is about
0.985 or so, which seems good to me. I'm not sure how to estimate the
possible error here, it might be interesting to try another test.

* feat: Calculate boundary area on saving Landscape (#210)

* feat: Add migrations for landscape boundary area (#211)

* fix: Change area_scalar to reflect units

* fix: Use GeoJSON default CRS for all boundaries

This is outlined in https://www.rfc-editor.org/rfc/rfc7946#section-4
. We should only support what the standard supports, do no need to do extra.

* feat: Add boundary area field to GraphQL schema (#210)

* fix: Add missing m2 to area storage (#210)

* feat: Add area in hectares to Landscape GraphQL Schema

* fix: Use default CRS instead of giving option

Because the current GeoJSON spec only uses a default CRS and removes
the ability to specify a custom CRS, we shouldn't support custom CRS either.

* feat: Add django-dirtyfields library

* feat: Only calculate landscape area if boundary has changed

* fix: Fix up migration conflict

* fix: Address PR comments

* fix: Calculate area for multiple GeoJSON features

I mistakenly thought that the landscape boundary was only a single
polygon, but this is not the case. The new function calculates area
for any collection of Polygon or MultiPolygon. Caveats: I didn't bother cover
GeometryCollection yet, and it doesn't take into account intersections.

* fix: Fix hectare conversion (#222)

There is 10,000 meters squared in a hectare, not 1,000

* fix: install GDAL in Docker image

* fix: remove unneeded ppa

* feat: Add changes to delete landscape profile photo (#223)

* feat: Add profile image fields to mutation
* feat: Add test for updating landscape profile image

* chore: Spanish taxonomy terms (#225)

Co-authored-by: David Code Howard <david@techmatters.org>
Co-authored-by: Paul Schreiber <paul@techmatters.org>