feat: Add option for reset/backup from different environment to Django admin (#212)
* feat: (WIP) Added entry type

* feat: Share links add mutation

* fix: Return link url if not file

* fix: Add resource type to add data entry

* feat: add Django script to backup DB

* feat: Add loadbackup script

* feat: Add option to upload backup file to S3 bucket

* feat: Add option to download backup from S3 bucket

* feat: Use boto3 instead of custom file copy

* feat: Address review comments

* feat: Add option to rewrite URLs on DB reload

* feat: Add more documentation on how to use backup script

* feat: Create script to run backend and S3 sync

This script does the following:

1) syncs the S3 buckets that need to be copied
2) backs up DB A to the backup S3 bucket
3) loads DB backup into DB B

It can be quite confusing, because the script is *not* designed to be
run in either environment A or B. You can think about it it as being
run in some environment C, which could just be your local machine. It
executes commands in environment A and B through the Render HTTP
API. It also uses the AWS cli to sync the S3 buckets.

The AWS cli needs to be installed in C. I provided a Dockerfile
as well here, but it will need to have the AWS cli credentials mounted
when it is run. It also needs two environment variables to be
provided.
    - $RENDER_TOKEN is the Render HTTP API Token
    - $URL_REWRITES_CONFIG_FILE is a file that should be stored in B
    and indicates the location of a config file that defines the URLs
    that need to be replaced in the database to point to the B
    environment S3 buckets. (This is documented elsewhere). The idea
    is this file can be provided as a "secret file" in the Render interface.

* feat: make configuration files for backup script

* Update backup/Dockerfile

* feat: make configuration files for backup script

* feat: Override the index page

This is mostly just to figure out how the template overrides
works. Here we are overriding the admin page index with our own
template, which just extends it to produce the same view. The next
step will be to add our custom button here.

* feat: backend routes for custom logic in Django admin

* feat: Add model for background tasks

This model will be used to keep track of jobs launched as threads that
we would like to keep track of. The thread will be able to set the
status and the frontend will be able to confirm it.

* feat: Run pseudo-backup in threads and store in DB

This is a simple way to run longer running tasks in the background. It
will work well for the uncommon restores. For now it just includes a
fake job that sleeps for 15 seconds.

* feat: Add logic to run backup/restore as backend task

This commit adds the three steps needed here:
1) sync s3 buckets
2) upload backup from "source" instance
3) download backup to "target" instance and restore

It is still missing a few important steps:
- need to provide a sample config file for the restore
- need to add the AWS cli to the Docker image, because we call the cli
directly. This is done because boto does not support the sync method.

* feat: Add missing configuration for reset

* fix: Address PR comments

* fix: Fix several bugs

* fix: Keep same user with the same password and user id after wipe

* fix: Keep user logged in during restore

* fix: Fix user id switch

Using data from staging, it became clear the user ID switch wasn't
working. This commit fixes this issue.

* chore: Add documentation for backup/reset

* fix: Address MR comments

* fix: Fix background task against main

Because I was working earlier with staging, I had worked on top of a
lot of different of different commits in the past. When I removed the
migration commits, it became necessary to move the migration pointers.

* fix: Remove non-existing migrations (for now)

* fix: Address review comments

* fix: Apply suggestions from code review

Co-authored-by: Paul Schreiber <paul@techmatters.org>

* fix: Apply further suggestions from code review

Co-authored-by: Paul Schreiber <paul@techmatters.org>

* fix: "Rebase" core migrations to remove branch

* fix: Fix changes to user ID after restore

* fix: Update migration data for loadbackup

Co-authored-by: Jose Buitron <jose.buitron@gmail.com>
Co-authored-by: Paul Schreiber <paul@techmatters.org>