fix: Allow recreating landscape (#229)
* fix: Add failing tests for bug

* fix: Change unique constraint to match safedelete docs

https://django-safedelete.readthedocs.io/en/latest/models.html#fields-uniqueness

Unfortunately this means that our error messages are slightly less
expressive, so I had to fix another test to remove a check for the
error message.

* fix: Add hacky fix to delete default groups on landscape deletion

For some reason the default landscape group is not deleted when the
corresponding landscape is deleted. I need to look more into why this
is the case. For the mean time this fixes the problem.

* fix: Delete default group on landscape deletion

The default group is like a secret 1-to-1 relation, but it is not
automatically deleted when the landscape is deleted. This commit fixes
that by overriding the model delete method to delete the default group
as well.

* chore: Add test to check that landscape can be recreated