fix: Report error name properly on test failure
The problem: for django-safe-delete to work properly, we cannot define
the constraints at a field level. The solution is to define a
constraint at the class level instead.
https://django-safedelete.readthedocs.io/en/latest/models.html#fields-uniqueness

This commit reworks the "BaseModel" class so that we can retrieve a
meaningful error message from the error. We assign the field name as
the error message for the table level Constraint, and then override
the validate method to associate the correct validation error to this field.