fix: Allow metaclass to override previously developed unique fields
There is a subclass (TaxonomyTerm) that overrode the unique
constraint declared by the Slug metaclass. This commit adds an option
to ignore the unique field declared in the parent class if we want to
create a different unique constraint in the child.

The drawback here is that the unique class may not be recognized by
our "error reporting" set up in the BaseModel. However because the
slug is not input by the user, that should be fine for now.