fix: Make data entry not unset when user deleted
The problem is that when the user is hard deleted, the created_by will
still remain pointing to a user who does not exist. Need to look into
if it's possible to fix this on a hard delete (probably) OR override
soft delete policy to reset the user id on delete.