fix: Fix error message for duplicate landscape name creation (#237)
* fix: Add hacky fix to delete default groups on landscape deletion

For some reason the default landscape group is not deleted when the
corresponding landscape is deleted. I need to look more into why this
is the case. For the mean time this fixes the problem.

* fix: Delete default group on landscape deletion

The default group is like a secret 1-to-1 relation, but it is not
automatically deleted when the landscape is deleted. This commit fixes
that by overriding the model delete method to delete the default group
as well.

* fix: Report error name properly on test failure

The problem: for django-safe-delete to work properly, we cannot define
the constraints at a field level. The solution is to define a
constraint at the class level instead.
https://django-safedelete.readthedocs.io/en/latest/models.html#fields-uniqueness

This commit reworks the "BaseModel" class so that we can retrieve a
meaningful error message from the error. We assign the field name as
the error message for the table level Constraint, and then override
the validate method to associate the correct validation error to this field.

* fix: Update metaclass to no longer use "_constraints" field

* fix: Allow metaclass to override previously developed unique fields

There is a subclass (TaxonomyTerm) that overrode the unique
constraint declared by the Slug metaclass. This commit adds an option
to ignore the unique field declared in the parent class if we want to
create a different unique constraint in the child.

The drawback here is that the unique class may not be recognized by
our "error reporting" set up in the BaseModel. However because the
slug is not input by the user, that should be fine for now.

* fix: Add missing migration to update dependencies

I had updated a previous migration (I think just to get some tests to
work). This commit does this properly by removing the constraints and
recreating them.

* chore: Add comments, clean up code

* fix: remove debug print() call

* fix: Remove unneeded "deleted_at" field alterations from migration

Co-authored-by: Paul Schreiber <paul@techmatters.org>