fix: Read OIDC ISS endpoint from config
Moodle (and I presume other services) rely on a file
/.well-known/openid-configuration/ . This is generated automatically
by the library django-oauth-toolkit based on the configuration values
we provide. Unfortunately inside the file it seems to generate
endpoints with the  http:// schema. It's a bit difficult to figure why
this is happening, so an easy fix is to provide the correct value
directly. Would be worth looking more into this in the future.