feature: oauth server 246 (#269)
* fix: Read OIDC ISS endpoint from config

Moodle (and I presume other services) rely on a file
/.well-known/openid-configuration/ . This is generated automatically
by the library django-oauth-toolkit based on the configuration values
we provide. Unfortunately inside the file it seems to generate
endpoints with the  http:// schema. It's a bit difficult to figure why
this is happening, so an easy fix is to provide the correct value
directly. Would be worth looking more into this in the future.

* fix: Stop displaying OAuth views on production server

django-oauth-toolkit comes with some views enabled that allow any
users to create OAuth applications that can use our server. We don't
want normal users to be able to this. This commit thus removes these
endpoints from the urls. Follows the example from https://django-oauth-toolkit.readthedocs.io/en/latest/tutorial/tutorial_02.html

* test: Add test to make sure userinfo returns email

Having some problems with the Kobo SSO, it seems like our email isn't
being picked up properly. This test is just to make sure that it is
really being produced!