feat: Add "restricted" membership type to groups
This new membership type is meant to accomodate groups that members
cannot join of their volition, but to which they must be added by a
manager. Project privacy can then be managed at the project level,
controlling which users can view and edit sites, and can be treated
orthogonally from the group membership type.