feature: Modify sites model to allow personal sites to be created (#511)
* feat: Create graphql mutation for site creation

* feat: Rename Site latitude and longitude fields

* chore: Rename soilproj app to project_management

* feat: Create graphql mutation for site creation

* feat: Save creator on Site during creation #374

This will be needed for Projects; we need to know what sites belong to
what users.

* feat: Add mutation to create new project

* feat: Create GraphQL endpoint to add site to project

This is a naive implementation that does not check if the user is a
member of the project, if they are administrator, etc. Next step will
be to add a test for these cases.

* feat: Restrict adding site to project to site creator

There is a bug with the base GraphQL cases that I will report. It
seems like the error message is not being returned as we want it to be.

* feat: Restrict adding sites to project to managers only

* chore: Fixup issues from Django app rename

Wow, that is kind of annoying to do. Hopefully all of the hanging
issues have been caught.

* fix: Fix up some errors caused by renaming project_management

* fix: Merge and fix up migrations

* fix: Add edit site mutation

This mutation can also be used to change the group the site belongs
to. Not sure if that is indeed what we want; perhaps this will be
split into a separate mutation as before.

* feat: Finish minimal project implementation

* fix: Squash mutations

* feat: Add "restricted" membership type to groups

This new membership type is meant to accomodate groups that members
cannot join of their volition, but to which they must be added by a
manager. Project privacy can then be managed at the project level,
controlling which users can view and edit sites, and can be treated
orthogonally from the group membership type.

* fix: Fix up and squash migrations

Squashed the migrations as the first one is not really valid by
itself (wrong unique name constraint).

* feat: Refactor add membership permissions

This commit just cleans up some of the logic for calculating if a user
is able to join a group or not.

* chore: Move project and site models into separate files

* feat: Associate site to project or user on creation

This is meant to support personal sites specifically.

* feat: Handle changing site owned by user to project