feat: Use MembershipList with Projects (#793)
* feat: Switch projects to use membership list; skip failing tests

* feat: Create mutation to update membership list of project

* feat: Restrict membership roles during creation

* feat: Add mutation for removing users from project

* test: Test permissions for deleting memberships

* test: Test user not member of project cannot update

* test: Test user can remove themself from project

* feat: Add mutation for updating membership role

* test: Add test for user not part of project updating role

* test: Reactivate skipped project tests

* feat: Move projects graphql into projects app

* refactor: Define and use project management collaboration roles

* chore: Make migration for project membership list

* fix: Enum types for ProjectMemberships, PR fixups

* feat: Use mixins to allow customizing MembershipNodes

* feat: Use enum for project update mutations

* fix: Restore missing membership_lists filter

This can be done be reusing the filterset class with
DjangoFilterConnectionField. It would be worthwhile checking if this
work can be encapsulated in a resuable function, as there is now a lot
of custom code whose function is not obvious.

* fix: Fix schema errors with ProjectMembership

The ProjectMembershipList was being a bunch of places where we did not
want it to be used, like Story Maps! I tried creating using some
Django proxy inheritance and that seems to have helped Graphene figure
things out.

* fix: Change Project model to point to proxy

* fix: Fix malfunctioning ProjectMembershipListNode

Somewhat confusing.

* fix: Fix project filter to correct memberships

Previously it was just returning all memberships for every project!

* chore: Reorder migrations