fix: Use ProjectMembershipNodes in Project mutation return types (#860)
This matches what the frontend expects, especially wrt to roles etc. I
had to redefine the utility method `get_member` to explicity use the
ProjectMembership object manager. There might be a more elegant way of
doing this with extending object managers, but I think that this will
do the trick for now.