feat: Ignore duplicate adding user to project (#844)
The plan is to use the project info in the redux store in the app to
determine if a user belongs to a project. This would create a corner
case where a user might be added to the same project twice. (Actually,
I think that would exist regardless). This commit just treats that
case a succesful operation and does nothing.