fix: Allow contributors to create sites (#906)
* fix: Allow contributors to create sites

* fix: Remove tests testing project settings

* chore: Add comment to project settings

Originally the project settings were supposed to give a normal member
the ability to add sites to project, etc. Since then we've changed the
project member role to viewers and contributors. I'm not removing
these settings here, because it might be used for other things in the
future, but I left a comment for anyone trying to figure this out in
the future.