fix: Create project soil settings on project creation (#1062)
Ran into a bug on the frontend when I tried creating a new
project. This commit makes sure that the project has a soil project
settings associated it. It also allows this to be disabled if we're
creating a project from somewhere where the project soil settings is
not necessary.