feat: Delete project data on a preset change (#1081)
* fix: Create project soil settings on project creation

Ran into a bug on the frontend when I tried creating a new
project. This commit makes sure that the project has a soil project
settings associated it. It also allows this to be disabled if we're
creating a project from somewhere where the project soil settings is
not necessary.

* feat: Apply changes on project preset change

* test: Add test to make sure depth dependent data preserved

* test: Add test to make sure depth dependent data deleted

* test: Test updating project preset to custom

* fix: Restrict project depth intervals on preset

* feat: Don't create project intervals on project soil creation

For projects that don't have the custom preset, the goal is to manage
these on the frontend.

* fix: Apply PR suggestions