refactor: move soil id perm code up to mutation
the implementation of the custom preset checks was causing confusion, because it was not obvious how `allowed_to_change_depth_interval` was meant to interact with other user role checks. since it was only used once, remove the method and extract the logic it contained to the interval mutation. this will yield greater conceptual clarity for authz vs. validation concerns.